Update followers every 12 hours;
After a user searches for followers, these are stored in the database to be accessed later.
However, a user can gain new followers on Twitter at any time, so we need to be fetching all new followers for saved users periodically every 12 hours.
This sets up a batch worker to hit the Twitter API every 12 hours to retrieve new followers and save them to the database.
If the follower already exists, don't save it.